Removed internal exports and adopted core package changes in @itwin/object-storage-minio package
