+++ storage/minio/src/frontend/MinioFrontendStorage.ts
@@ -5,14 +5,13 @@
 import {
   FrontendConfigUploadInput,
   FrontendUrlUploadInput,
-  instanceOfUrlTransferInput,
-  metadataToHeaders,
-  uploadToUrlFrontend,
 } from "@itwin/object-storage-core/lib/frontend";
+import { instanceOfUrlTransferInput } from "@itwin/object-storage-core/lib/common/internal";
 import {
   FrontendS3ClientWrapperFactory,
   S3FrontendStorage,
 } from "@itwin/object-storage-s3/lib/frontend";
+import { handleMinioUrlUploadFrontend } from "./internal";
 
 export class MinioFrontendStorage extends S3FrontendStorage {
   public constructor(clientWrapperFactory: FrontendS3ClientWrapperFactory) {
@@ -26,14 +25,4 @@ export class MinioFrontendStorage extends S3FrontendStorage {
       return handleMinioUrlUploadFrontend(input);
     else return super.upload(input);
   }
-}
-
-export async function handleMinioUrlUploadFrontend(
-  input: FrontendUrlUploadInput
-): Promise<void> {
-  const { data, metadata, url } = input;
-  const headers = metadata
-    ? metadataToHeaders(metadata, "x-amz-meta-")
-    : undefined;
-  return uploadToUrlFrontend(url, data, headers);
 }
