+++ storage/minio/src/frontend/internal/Helpers.ts
@@ -3,6 +3,9 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import { S3Client } from "@aws-sdk/client-s3";
+import { metadataToHeaders } from "@itwin/object-storage-core/lib/common/internal";
+import { FrontendUrlUploadInput } from "@itwin/object-storage-core/lib/frontend";
+import { uploadToUrlFrontend } from "@itwin/object-storage-core/lib/frontend/internal";
 
 export function createMinioS3ClientFrontend(config: {
   baseUrl: string;
@@ -31,4 +34,14 @@ export function createMinioS3ClientFrontend(config: {
     },
     forcePathStyle: true,
   });
+}
+
+export async function handleMinioUrlUploadFrontend(
+  input: FrontendUrlUploadInput
+): Promise<void> {
+  const { data, metadata, url } = input;
+  const headers = metadata
+    ? metadataToHeaders(metadata, "x-amz-meta-")
+    : undefined;
+  return uploadToUrlFrontend(url, data, headers);
 }
