+++ storage/minio/src/frontend/wrappers/FrontendMinioS3ClientFactory.ts
@@ -6,11 +6,11 @@ import { injectable } from "inversify";
 
 import { TransferConfig } from "@itwin/object-storage-core/lib/frontend";
 import {
-  assertS3TransferConfig,
   FrontendS3ClientWrapper,
 } from "@itwin/object-storage-s3/lib/frontend";
+import { assertS3TransferConfig } from "@itwin/object-storage-s3/lib/common/internal";
 
-import { createMinioS3ClientFrontend } from "../Helpers";
+import { createMinioS3ClientFrontend } from "../internal/Helpers";
 
 @injectable()
 export class FrontendMinioS3ClientWrapperFactory {
