+++ storage/minio/src/server/MinioPresignedUrlProvider.ts
@@ -6,10 +6,10 @@ import { inject, injectable } from "inversify";
 import { Client } from "minio";
 
 import {
-  buildObjectKey,
   ObjectReference,
   PresignedUrlProvider,
 } from "@itwin/object-storage-core";
+import { buildObjectKey } from "@itwin/object-storage-core/lib/common/internal";
 import { Types } from "@itwin/object-storage-s3";
 
 @injectable()
